Feature Engineering
Feature engineering is the process of selecting, extracting, and transforming raw data into features that can be used to improve the performance of machine learning models. It involves identifying and creating the most relevant and meaningful features that can accurately represent the underlying patterns and relationships in the data. 
This process can include techniques such as data cleaning, data normalization, feature selection, feature scaling, and feature construction, among others.
 The goal of feature engineering is to improve the quality of the data fed into a machine learning model, which can ultimately lead to better model accuracy and performance.